requirements.txt
numpy>=1.21.1
pandas>=1.3.5
PyPDF3>=1.0.5
setuptools>=49.2.1
PyYAML>=6.0
scipy>=1.7.3
coloredlogs>=15.0.1
fpdf2>=2.4.6

pytreemap>=0.4
statsmodels>=0.13.1
patsy>=0.5.2
pydream>=2.0.0
sklearn>=0.0
scikit-learn>=1.0.2
fastdtw>=0.3.4